use 0 as default if no build introduced available
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonShared/trunk@67077 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ DatasetPresenter/src/java/org/apidb/apicommon/datasetPresenter/DatasetPresenterSetLoader.java
@@ -389,7 +389,7 @@ public class DatasetPresenterSetLoader {
     Integer buildNumberIntroduced = datasetPresenter.getBuildNumberIntroduced();
 
     if(buildNumberIntroduced == null) {
-        buildNumberIntroduced = new Integer(1);
+        buildNumberIntroduced = new Integer(0);
     }
 
     stmt.setInt(i++, buildNumberIntroduced.intValue());
